refactor: modify token setting logic
