+++ src/utils/authApi.ts
@@ -1,46 +1,21 @@
 "use server";
 import returnFetch from "return-fetch";
-import { getCookie, setCookie } from "cookies-next";
 import { cookies } from "next/headers";
-import { camelCase } from "lodash";
+import { getRefreshToken } from "./common";
 
 const baseURL = process.env.NEXT_PUBLIC_BACKEND;
-
-const setRefreshCookie = (cookieStr: string) => {
-  let refreshTokenValue = "";
-  let cookieOptions = {};
-  cookieStr.split(";").map((item) => {
-    const cookieItem = item.trim().split("=");
-    if (cookieItem.includes("Refresh")) {
-      refreshTokenValue = cookieItem[1];
-    } else {
-      const optionName = camelCase(cookieItem[0]);
-      const optionValue = cookieItem[1] ?? true;
-      cookieOptions = {
-        ...cookieOptions,
-        [optionName]: optionValue,
-      };
-    }
-  });
-
-  console.log("refreshTokenValue", refreshTokenValue);
-  console.log("cookieOptions", cookieOptions);
-
-  if (refreshTokenValue) {
-    cookies().set("Refresh", refreshTokenValue, cookieOptions);
-  }
-};
-
 const authApi = returnFetch({
   baseUrl: baseURL,
   interceptors: {
     request: async (requestArgs) => {
-      const accessToken = getCookie("Access", { cookies });
+      const cookieStore = cookies();
+      const accessToken = cookieStore.get("Access");
+
       if (requestArgs[1] && accessToken) {
         requestArgs[1].headers = {
           "Content-Type": "application/json",
           ...requestArgs[1]?.headers,
-          Authorization: `Bearer ${accessToken}`,
+          Authorization: `Bearer ${accessToken.value}`,
         };
       }
 
@@ -54,15 +29,16 @@ const authApi = returnFetch({
 
       console.log("not authorized, trying to get refresh cookie..");
 
-      const userId = getCookie("user_id", { cookies });
-      const refreshToken = getCookie("Refresh", { cookies });
+      const cookieStore = cookies();
+      const userId = cookieStore.get("user_id");
+      const refreshToken = cookieStore.get("Refresh");
       if (userId && refreshToken) {
         const tokenResponse = await fetch(`${baseURL}/api/user/token-reissue`, {
           method: "POST",
-          body: JSON.stringify({ userId }),
+          body: JSON.stringify({ userId: userId.value }),
           headers: {
             "Content-Type": "application/json",
-            Cookie: `Refresh=${refreshToken}`,
+            Cookie: `Refresh=${refreshToken.value}`,
           },
           credentials: "include",
         });
@@ -70,11 +46,13 @@ const authApi = returnFetch({
         if (tokenResponse.ok) {
           const { headers } = tokenResponse;
           const accessToken = headers.get("Authorization");
-          setCookie("Access", accessToken, { cookies });
+          const setCookieHeader = headers.get("Set-Cookie");
 
-          const refreshToken = headers.get("Set-Cookie");
-          if (refreshToken) {
-            setRefreshCookie(refreshToken);
+          if (accessToken && setCookieHeader) {
+            const { token, options } = getRefreshToken(setCookieHeader);
+
+            cookieStore.set("Access", accessToken);
+            cookieStore.set("Refresh", token, options);
           }
         } else {
           throw Error("failed to refresh cookie");
